requirements.txt
# FYP Job Matching Backend - Production Requirements
# Install with: pip install -r requirements.txt

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
supabase==2.3.0

# Environment Management
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.0

# HTTP Client
httpx==0.27.0

# Production Server
gunicorn==21.2.0

# Development Tools (Optional - remove for production)
# black==24.1.0
# flake8==7.0.0
# pytest==8.0.0